fix(review): Move StrawberryTypeFromPydantic to new file. Add example to RELEASE.md.
