+++ strawberry/experimental/pydantic/object_type.py
@@ -13,13 +13,12 @@ from typing import (
     Optional,
     Sequence,
     Type,
-    TypeVar,
     cast,
 )
 
 from pydantic import BaseModel
 from pydantic.fields import ModelField
-from typing_extensions import Literal, Protocol
+from typing_extensions import Literal
 
 from graphql import GraphQLResolveInfo
 
@@ -87,31 +86,10 @@ def get_type_for_field(field: ModelField):
 
 
 if TYPE_CHECKING:
-    PydanticModel = TypeVar("PydanticModel", bound=BaseModel)
-
-    class StrawberryTypeFromPydantic(Protocol[PydanticModel]):
-        """This class does not exist in runtime.
-        It only makes the methods below visible for IDEs"""
-
-        def __init__(self, **kwargs):
-            ...
-
-        @staticmethod
-        def from_pydantic(
-            instance: PydanticModel, extra: Dict[str, Any] = None
-        ) -> PydanticModel:
-            ...
-
-        def to_pydantic(self) -> PydanticModel:
-            ...
-
-        @property
-        def _type_definition(self) -> TypeDefinition:
-            ...
-
-        @property
-        def _pydantic_type(self) -> PydanticModel:
-            ...
+    from strawberry.experimental.pydantic.conversion_types import (
+        PydanticModel,
+        StrawberryTypeFromPydantic,
+    )
 
 
 def type(
